Add composite action to create upstream tag
The release_bot.py contains several steps, which were previously part of
our release script in osbuild/maintainer-tools. The script here is much
shorter though.
Things that were dropped were the sanity checks, because we can presume
that the git repository at hand is sane, that the git remote is well
configured etc.
Furthermore all interactivity of the original script was removed. Even
though I had recently added a --no-interactive mode to the original
script, I just felt this would make the code harder to read.

The rationale for making this a composite action instead of a standalone
repository was that this way each component can define it's own
automated schedule and that there would be some representation in the
GitHub Actions for the component of the tag being created.