Fix detection of no commits since the last version
Previously, the code used `split('\n')` on commands, which were listing
commits since the last release. However, splitting an empty string with
a specified separator returns [''], which fooled the code to behave as
if there was one commit since the last release.

As a result, the code would push a new release tag, even if there were
no new commits.

Fix this by using `splitlines()`, which returns an empty list when used
on empty string.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>