create-tag: fix corner cases when bot creates / approves a PR
We relied on the ability to get reviewers full name to create the tag
message. However, bots (e.g. github-action) in general don't have this
property set. This resulted in the reviewers list to contain `None` as
its item, which consequently resulted in a failure of the script.

Modify the script to use the author's and / or reviewer's login as
their name, in case getting their full name fails.

If the list of reviewers of a PR is empty (e.g. admin-merged without any
approval), append "Nobody" to the list of reviewers.

Cover the special case of automated dependency updates (e.g. in
osbuild/images repo) and in such case don't list author an reviewers in
the tag message, but a custom message.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>
Co-authored-by: Tomáš Hozza <thozza@redhat.com>