Always try to determine the latest version from tags
The latest version from tags was previously not determined if the
version to update to was provided as an argument to the script. This
behavior then impacted the collection of PRs to put into the changelog.

Since the script does not support making the first "dot" release
automatically, so one has to provide the version explicitly to the
script.

This change solves this situation by making the tag message sane and
complete.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>