Allow specifying the remote to push the tag to
While defaulting to pushing to 'origin' makes sense when the script is
run inside the GH action, it makes less sense when running it locally
(i.e. to create the first "dot" release tag). Many people have the
'origin' remote set up as read-only, so pushing to it won't work.
Therefore add a new CLI option to allow specifying the remote to push
the tag to.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>