create-tag: use dashes instead of asterisks
When creating a new tag our description uses `*`-asterisks to denote
an unordered list. This clashes with the format used in RPM spec files
where asterisks can have special meaning.

In some corner cases this leads to automatic tools not being able to
parse the changelogs as we embed them.

This is a very minor change since `-`-dash is also a valid unordered
list marker in markdown.

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>