release-action: Add changelog in Slack threaded message
Instead of simply posting the announcement of a new version in the
Slack channel, let's add a threaded response that contains the whole
release changelog for convenience. This way, people from other teams can
easily look up the changes that made it into a particular release
without leaving Slack / going to GitHub.